[mod, add, del] fix example project display and scroll in dialog
